Fix for the "More at..." link to correct lineId
The sourceUrl link which controls the "More at tfl.gov.uk" href is now
fixed to link to the correct anchor for the line in question. This will
open the further information accordion and zoom the SVG map to the
correct line. (There appears to be an issue with the lineIds being incorrect on the
map  on tfl.gov.uk at the moment?)